requirements.txt
pydantic>=2.7.3
pydantic[email]>=2.7.3
email-validator>=2.1.0
weasyprint>=64.0
Pillow>=11.0.0
pybars3>=0.9.7
pdf2image>=1.17.0
python-dateutil>=2.8.2 